Created a separate file for the class InteractiveBeamParserState
